+++ src/Symfony/Component/String/AbstractString.php
@@ -438,6 +438,13 @@ abstract class AbstractString implements \Stringable, \JsonSerializable
         return $this->snake()->replace('_', '-');
     }
 
+    public function pascal(): static
+    {
+        $this->string = ucfirst($this->camel());
+
+        return $this;
+    }
+
     abstract public function splice(string $replacement, int $start = 0, ?int $length = null): static;
 
     /**