+++ src/Symfony/Component/String/Tests/AbstractAsciiTestCase.php
@@ -1119,6 +1119,26 @@ abstract class AbstractAsciiTestCase extends TestCase
     }
 
     /**
+     * @testWith ["", ""]
+     *           ["XY", "x_y"]
+     *           ["XuYo", "xu_yo"]
+     *           ["SymfonyIsGreat", "symfony_is_great"]
+     *           ["Symfony5IsGreat", "symfony_5_is_great"]
+     *           ["SymfonyIsGreat", "Symfony is great"]
+     *           ["SYMFONYISGREAT", "SYMFONY_IS_GREAT"]
+     *           ["SymfonyIsAGreatFramework", "Symfony is a great framework"]
+     *           ["SymfonyIsGREAT", "*Symfony* is GREAT!!"]
+     *           ["SYMFONY", "SYMFONY"]
+     */
+    public function testPascal(string $expectedString, string $origin)
+    {
+        $instance = static::createFromString($origin)->pascal();
+
+        $this->assertEquals(static::createFromString($expectedString), $instance);
+        $this->assertNotSame($origin, $instance, 'Strings should be immutable');
+    }
+
+    /**
      * @dataProvider provideStartsWith
      */
     public function testStartsWith(bool $expected, string $origin, $prefix, ?int $form = null)